Computes census of imaginary errors

Computes census of imaginary errors

Usage

count_imaginary_census(x, counter_type = 0L)

Arguments

x

An object of class barry_graph.

counter_type

An integer indicating the type of census to compute (see details).

Details

We can also separate the counts as a function of whether the perceiver is looking into all ties, only ties including them, or only ties not including them. This is controlled by the counter_type argument:

0: All ties

1: Only ties including the perceiver

2: Only ties not including the perceiver

There are ten (10) values:

(01) Accurate null

(02) Partial false positive (null)

(03) Complete false positive (null)

(04) Partial false negative (assym)

(05) Accurate assym

(06) Mixed assym

(07) Partial false positive (assym)

(08) Complete false negative (full)

(09) Partial false negative (full)

(10) Accurate full