Binary Array Graph

Binary Array Graph

Usage

new_barry_graph(x, ...)

# S3 method for class 'matrix'
new_barry_graph(x, n, ...)

# S3 method for class 'list'
new_barry_graph(x, ...)

Arguments

x

Either a matrix or a list of matrices.

...

Currently ignored.

n

Integer. The size of the original network.

Details

When x is a matrix, it is assumed that it will be a block diagonal matrix, with the first block corresponding to the reference (true) network.

If x is a list, the first matrix is assumed to be the reference (true) network.